Use DropdownActions component
